+++ app/Http/Controllers/AccountController.php
@@ -72,7 +72,6 @@ class AccountController extends Controller
         return Redirect::back()->with(['success' => 'Your details have successfully been changed']);
     }
 
-
     /**
      * @author EB
      * @param Request $request
