+++ tests/Controllers/AccountControllerTest.php
@@ -1,0 +1,49 @@
+<?php
+/*
+ * This file is part of the PayBreak/basket package.
+ *
+ * (c) PayBreak <dev@paybreak.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use App\User;
+use App\Http\Controllers;
+
+class AccountControllerTest extends TestCase
+{
+    /**
+     * @author WN
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        Artisan::call('migrate');
+        Artisan::call('db:seed', ['--class' => 'DBSeeder']);
+
+        $user = User::find(1);
+        $this->be($user);
+    }
+
+    /**
+     * @author WN
+     */
+    public function testShow()
+    {
+        // Test page gives 200 response
+        $this->visit('/account')
+            ->seeStatusCode(200);
+    }
+
+    /**
+     * @author WN
+     */
+    public function testEdit()
+    {
+        // Test page gives 200 response
+        $this->visit('/account/edit')
+            ->seeStatusCode(200);
+    }
+}
